Improve ESP32 HAL (EEPROM, watchdog) (#16228)
