+++ platformio.ini
@@ -694,14 +694,16 @@ monitor_speed = 250000
 [env:esp32]
 platform      = espressif32
 board         = esp32dev
-upload_speed  = 115200
-monitor_speed = 115200
-upload_port   = /dev/ttyUSB0
+build_flags   = ${common.build_flags} -DCORE_DEBUG_LEVEL=0
 lib_deps      =
   AsyncTCP=https://github.com/me-no-dev/AsyncTCP/archive/master.zip
   ESPAsyncWebServer=https://github.com/me-no-dev/ESPAsyncWebServer/archive/master.zip
 lib_ignore    = LiquidCrystal, LiquidTWI2, SailfishLCD, SailfishRGB_LED
 src_filter    = ${common.default_src_filter} +<src/HAL/HAL_ESP32>
+upload_speed  = 115200
+monitor_speed = 250000
+#upload_port   = marlinesp.local
+#board_build.flash_mode = qio
 
 #
 # Native
